+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 ### Fixes
 - Crash if removing a character from main screen editor
 - Crash if leaving or restart a scene if you were not in an editor (tries to stop it while it's null)
+- Packet for removing template characters in fabric not properly registered